Move util package from domain to project root and update imports
