+++ app/src/main/java/ir/miare/androidcodechallenge/domain/repository/PlayerRepository.kt
@@ -2,7 +2,7 @@ package ir.miare.androidcodechallenge.domain.repository
 
 import androidx.paging.PagingData
 import ir.miare.androidcodechallenge.domain.model.Player
-import ir.miare.androidcodechallenge.domain.util.SortOption
+import ir.miare.androidcodechallenge.util.SortOption
 import kotlinx.coroutines.flow.Flow
 
 interface PlayerRepository {
